// src/app/interfaces/records-audio-response.interface.ts
export interface RecordsAudioResponse {
  readonly AlbumTitle: string;
  readonly CoverImageSource: string;
  readonly CoverImageAltText: string;
  readonly Description: string;
  readonly AlbumName: string;
  readonly TrackSource: string;
  readonly RecordTitle: string;
  readonly SoloistName: string | null;
}